+++ CHANGELOG.md
@@ -38,11 +38,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - When a tree viewer is refreshed using JSON from a file with a different name, the title of the docking form that the user sees doesn't change to reflect the new file. For example, a tree viewer is opened up for `foo.json` and then refreshed with a buffer named `bar.json`, and the title of the docking form still reads `Json Tree View for foo.json`.
 	- This is also true if a file with a tree viewer is renamed, e.g., the file `foo.json` is renamed to `bar.json`, but the tree viewer still says `Json Tree View for foo.json`.
 
-## [5.3.0] - (UNRELEASED) 2023-MM-DD
+## [5.3.0] - 2023-06-10
 
 ### Added
 
-1. Optional `sort_by_count` argument for [`value_counts` function](/docs/RemesPath.md#non-vectorized-functions), which sorts subarrays by count if true.
+1. New [error form](/docs/README.md#error-form-and-status-bar) for viewing syntax errors.
+2. When JSON is parsed, the [document type status bar section changes](/docs/README.md#error-form-and-status-bar) to reflect the document's level of compliance with the JSON standard.
+2. Optional `sort_by_count` argument for [`value_counts` function](/docs/RemesPath.md#non-vectorized-functions), which sorts subarrays by count if true.
 
 ## [5.2.0] - 2023-06-04